Move photoUpload files and update state successfully
