+++ client/app/user/photoUpload/PhotoUploadContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import PhotoUpload from './PhotoUpload';
-import * as UserActions from './UserActions';
+import * as UserActions from '../UserActions';
 
 const mapStateToProps = (state) => (
   {
